api/requirements.txt
Flask
flask_restful
psycopg2
flask_sqlalchemy
flask_migrate
marshmallow
